+++ Moonshot/GridView.swift
@@ -1,0 +1,50 @@
+//
+//  GridView.swift
+//  Moonshot
+//
+//  Created by Razvan Pricop on 24.10.24.
+//
+
+import SwiftUI
+
+struct GridView: View {
+    let astronauts: [String: Astronaut]
+    let missions: [Mission]
+    
+    var body: some View {
+        List(missions) { mission in
+            NavigationLink {
+                MissionView(mission: mission, astronauts: astronauts)
+            } label: {
+                VStack {
+                    Image(mission.image)
+                        .resizable()
+                        .scaledToFit()
+                        .frame(width: 100, height: 100)
+                        .padding()
+                    
+                    VStack {
+                        Text(mission.displayName)
+                            .font(.headline)
+                            .foregroundStyle(.white)
+                        Text(mission.formattedLaunchDate)
+                            .font(.caption)
+                            .foregroundStyle(.white.opacity(0.5))
+                    }
+                    .padding(.vertical)
+                    .frame(maxWidth: .infinity)
+                    .background(.lightBackground)
+                }
+                .clipShape(.rect(cornerRadius: 10))
+                .overlay(RoundedRectangle(cornerRadius: 10).stroke(.lightBackground))
+            }
+        }
+    }
+}
+
+#Preview {
+    let missions: [Mission] = Bundle.main.decode("missions.json")
+    let astronauts: [String: Astronaut] = Bundle.main.decode("astronauts.json")
+    
+    GridView(astronauts: astronauts, missions: missions)
+}
