+++ Moonshot/ListView.swift
@@ -1,0 +1,57 @@
+//
+//  ListView.swift
+//  Moonshot
+//
+//  Created by Razvan Pricop on 24.10.24.
+//
+
+import SwiftUI
+
+struct ListView: View {
+    let astronauts: [String: Astronaut]
+    let missions: [Mission]
+    let columns = [GridItem(.adaptive(minimum: 150))]
+    
+    var body: some View {
+        
+        ScrollView {
+            LazyVGrid(columns: columns) {
+                ForEach(missions) { mission in
+                    NavigationLink {
+                        MissionView(mission: mission, astronauts: astronauts)
+                    } label: {
+                        VStack {
+                            Image(mission.image)
+                                .resizable()
+                                .scaledToFit()
+                                .frame(width: 100, height: 100)
+                                .padding()
+                            
+                            VStack {
+                                Text(mission.displayName)
+                                    .font(.headline)
+                                    .foregroundStyle(.white)
+                                Text(mission.formattedLaunchDate)
+                                    .font(.caption)
+                                    .foregroundStyle(.white.opacity(0.5))
+                            }
+                            .padding(.vertical)
+                            .frame(maxWidth: .infinity)
+                            .background(.lightBackground)
+                        }
+                        .clipShape(.rect(cornerRadius: 10))
+                        .overlay(RoundedRectangle(cornerRadius: 10).stroke(.lightBackground))
+                    }
+                }
+            }
+            .padding([.horizontal, .bottom])
+        }
+    }
+}
+
+#Preview {
+    let missions: [Mission] = Bundle.main.decode("missions.json")
+    let astronauts: [String: Astronaut] = Bundle.main.decode("astronauts.json")
+    
+    ListView(astronauts: astronauts, missions: missions)
+}
