+++ Moonshot/MissionView.swift
@@ -8,25 +8,8 @@
 import SwiftUI
 
 struct MissionView: View {
-    struct CrewMember {
-        let role: String
-        let astronaut: Astronaut
-    }
-    
     let mission: Mission
-    let crew: [CrewMember]
-    
-    init(mission: Mission, astronauts: [String: Astronaut]) {
-        self.mission = mission
-        
-        self.crew = mission.crew.map { member in
-            if let astronaut = astronauts[member.name] {
-                return CrewMember(role: member.role, astronaut: astronaut)
-            } else {
-                fatalError("Could not find astronaut for \(member.name)")
-            }
-        }
-    }
+    let astronauts: [String: Astronaut]
     
     var body: some View {
         ScrollView {
@@ -39,52 +22,28 @@ struct MissionView: View {
                     }
                     .padding(.top)
                 
-                VStack(alignment: .leading) {
-                    Text("Mission Highlights")
-                        .font(.title.bold())
-                        .padding(.bottom, 5)
-                    
+                Text(mission.formattedLaunchDate)
+                    .padding(.bottom)
+                
+                Rectangle()
+                    .frame(height: 2)
+                    .foregroundStyle(.lightBackground)
+                    .padding(.horizontal)
+                
+                MissionHighlightsView(mission: mission)
+                    .padding(.horizontal)
+            
+                VStack {
                     Rectangle()
                         .frame(height: 2)
                         .foregroundStyle(.lightBackground)
-                    
-                    Text(mission.description)
-                    
-                    Rectangle()
-                        .frame(height: 2)
-                        .foregroundStyle(.lightBackground)
+                        .padding(.horizontal)
                     
                     Text("Crew")
                         .font(.title.bold())
                         .padding(.bottom, 5)
-                }
-                .padding(.horizontal)
-                
-                ScrollView(.horizontal, showsIndicators: false) {
-                    HStack {
-                        ForEach(crew, id: \.role) { crewMember in
-                            NavigationLink {
-                                AstronautView(astronaut: crewMember.astronaut)
-                            } label: {
-                                HStack {
-                                    Image(crewMember.astronaut.id)
-                                        .resizable()
-                                        .frame(width: 104, height: 72)
-                                        .clipShape(.capsule)
-                                        .overlay(Capsule().strokeBorder(.white, lineWidth: 1))
-                                    
-                                    VStack(alignment: .leading) {
-                                        Text(crewMember.astronaut.name)
-                                            .foregroundStyle(.white)
-                                            .font(.headline)
-                                        Text(crewMember.role)
-                                            .foregroundStyle(.white.opacity(0.5))
-                                    }
-                                }
-                                .padding(.horizontal)
-                            }
-                        }
-                    }
+                    
+                    CrewCarousel(mission: mission, astronauts: astronauts)
                 }
             }
             .padding(.bottom)
@@ -99,6 +58,6 @@ struct MissionView: View {
     let missions: [Mission] = Bundle.main.decode("missions.json")
     let astronauts: [String: Astronaut] = Bundle.main.decode("astronauts.json")
     
-    MissionView(mission: missions[0], astronauts: astronauts)
+    MissionView(mission: missions[1], astronauts: astronauts)
         .preferredColorScheme(.dark)
 }
